feat phpmyadminの長年のエラーは型の符号の有無だったのかもしれない。

+++ database/migrations/2020_09_27_183450_create_followers_table.php
@@ -18,8 +18,8 @@ class CreateFollowersTable extends Migration
              * 自分がフォローしているユーザーのあいさつを表示するとき
              * 自分がfollowing_id フォローした相手がfollowed_id
              */
-            $table->bigInteger('following_id')->comment('フォローしているユーザID');
-            $table->bigInteger('followed_id')->comment('フォローされているユーザID');
+            $table->unsignedInteger('following_id')->comment('フォローしているユーザID');
+            $table->unsignedInteger('followed_id')->comment('フォローされているユーザID');
             $table->timestamps();
 
             //インデックスで高速化
