+++ database/migrations/2020_09_27_183941_create_direct_messages_table.php
@@ -15,8 +15,8 @@ class CreateDirectMessagesTable extends Migration
     {
         Schema::create('direct_messages', function (Blueprint $table) {
             $table->increments('id');
-            $table->bigInteger('send_user_id')->comment('送信者');
-            $table->bigInteger('receive_user_id')->comment('受信者');
+            $table->unsignedInteger('send_user_id')->comment('送信者');
+            $table->unsignedInteger('receive_user_id')->comment('受信者');
             $table->text('message')->comment('メッセージ');
             $table->timestamps();
